AG-9869 - Fix RadarSeries resize.
